requirements.txt
Flask
python-dotenv
google-generativeai
gunicorn
you can generate your own gemini api key